refactor(preset): move presets endpoint implementation from "project" to "module" bounded context
