+++ src/main/java/tech/jhipster/lite/module/application/JHipsterModulesApplicationService.java
@@ -2,18 +2,13 @@ package tech.jhipster.lite.module.application;
 
 import java.util.Collection;
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.module.domain.GeneratedProjectRepository;
-import tech.jhipster.lite.module.domain.JHipsterModuleApplied;
-import tech.jhipster.lite.module.domain.JHipsterModuleEvents;
-import tech.jhipster.lite.module.domain.JHipsterModuleToApply;
-import tech.jhipster.lite.module.domain.JHipsterModulesApplyer;
-import tech.jhipster.lite.module.domain.JHipsterModulesRepository;
-import tech.jhipster.lite.module.domain.JHipsterModulesToApply;
+import tech.jhipster.lite.module.domain.*;
 import tech.jhipster.lite.module.domain.git.GitRepository;
 import tech.jhipster.lite.module.domain.javabuild.ProjectJavaBuildToolRepository;
 import tech.jhipster.lite.module.domain.javadependency.JavaDependenciesVersionsRepository;
 import tech.jhipster.lite.module.domain.javadependency.ProjectJavaDependenciesRepository;
 import tech.jhipster.lite.module.domain.landscape.JHipsterLandscape;
+import tech.jhipster.lite.module.domain.preset.Preset;
 import tech.jhipster.lite.module.domain.resource.JHipsterModulesResources;
 
 @Service
@@ -22,6 +17,7 @@ public class JHipsterModulesApplicationService {
   private final JHipsterModuleEvents events;
   private final JHipsterModulesRepository modules;
   private final JHipsterModulesApplyer applyer;
+  private final JHipsterPresetRepository preset;
 
   public JHipsterModulesApplicationService(
     JHipsterModuleEvents events,
@@ -30,10 +26,12 @@ public class JHipsterModulesApplicationService {
     ProjectJavaDependenciesRepository projectDependencies,
     ProjectJavaBuildToolRepository javaBuildTools,
     GitRepository git,
-    GeneratedProjectRepository generatedProject
+    GeneratedProjectRepository generatedProject,
+    JHipsterPresetRepository preset
   ) {
     this.events = events;
     this.modules = modules;
+    this.preset = preset;
 
     applyer = new JHipsterModulesApplyer(modules, currentVersions, projectDependencies, javaBuildTools, git, generatedProject);
   }
@@ -56,5 +54,9 @@ public class JHipsterModulesApplicationService {
 
   public JHipsterLandscape landscape() {
     return modules.landscape();
+  }
+
+  public Collection<Preset> getPresets() {
+    return preset.getPresets();
   }
 }
