+++ src/main/java/tech/jhipster/lite/module/domain/JHipsterPresetRepository.java
@@ -1,7 +1,7 @@
 package tech.jhipster.lite.module.domain;
 
 import java.util.Collection;
-import tech.jhipster.lite.project.domain.preset.Preset;
+import tech.jhipster.lite.module.domain.preset.Preset;
 
 public interface JHipsterPresetRepository {
   Collection<Preset> getPresets();
