+++ src/main/java/tech/jhipster/lite/module/domain/preset/Preset.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.project.domain.preset;
+package tech.jhipster.lite.module.domain.preset;
 
 import tech.jhipster.lite.project.domain.ModulesSlugs;
 import tech.jhipster.lite.shared.error.domain.Assert;
