+++ src/main/java/tech/jhipster/lite/module/infrastructure/primary/RestModuleToApply.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.project.infrastructure.primary;
+package tech.jhipster.lite.module.infrastructure.primary;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import tech.jhipster.lite.project.domain.ModuleSlug;
