+++ src/main/java/tech/jhipster/lite/project/application/ProjectsApplicationService.java
@@ -1,6 +1,5 @@
 package tech.jhipster.lite.project.application;
 
-import java.util.Collection;
 import org.springframework.stereotype.Service;
 import tech.jhipster.lite.project.domain.ProjectPath;
 import tech.jhipster.lite.project.domain.ProjectsRepository;
@@ -9,7 +8,6 @@ import tech.jhipster.lite.project.domain.download.ProjectsDownloader;
 import tech.jhipster.lite.project.domain.history.ProjectActionToAppend;
 import tech.jhipster.lite.project.domain.history.ProjectActionsAppender;
 import tech.jhipster.lite.project.domain.history.ProjectHistory;
-import tech.jhipster.lite.project.domain.preset.Preset;
 import tech.jhipster.lite.shared.projectfolder.domain.ProjectFolder;
 
 @Service
@@ -40,9 +38,5 @@ public class ProjectsApplicationService {
 
   public ProjectHistory getHistory(ProjectPath path) {
     return projects.getHistory(path);
-  }
-
-  public Collection<Preset> getPresets() {
-    return projects.getPresets();
   }
 }
