+++ src/main/java/tech/jhipster/lite/project/domain/ProjectsRepository.java
@@ -1,10 +1,8 @@
 package tech.jhipster.lite.project.domain;
 
-import java.util.Collection;
 import java.util.Optional;
 import tech.jhipster.lite.project.domain.download.Project;
 import tech.jhipster.lite.project.domain.history.ProjectHistory;
-import tech.jhipster.lite.project.domain.preset.Preset;
 
 public interface ProjectsRepository {
   void format(ProjectPath path);
@@ -14,6 +12,4 @@ public interface ProjectsRepository {
   void save(ProjectHistory history);
 
   ProjectHistory getHistory(ProjectPath path);
-
-  Collection<Preset> getPresets();
 }
