+++ src/main/java/tech/jhipster/lite/project/infrastructure/secondary/FileSystemProjectsRepository.java
@@ -6,17 +6,12 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Collection;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
-import tech.jhipster.lite.module.domain.ProjectFiles;
 import tech.jhipster.lite.project.domain.ProjectPath;
 import tech.jhipster.lite.project.domain.ProjectsRepository;
 import tech.jhipster.lite.project.domain.download.Project;
 import tech.jhipster.lite.project.domain.history.ProjectHistory;
-import tech.jhipster.lite.project.domain.preset.Preset;
-import tech.jhipster.lite.project.domain.preset.PresetName;
 import tech.jhipster.lite.shared.error.domain.Assert;
 import tech.jhipster.lite.shared.error.domain.GeneratorException;
 
@@ -27,25 +22,15 @@ class FileSystemProjectsRepository implements ProjectsRepository {
 
   private static final String HISTORY_FOLDER = ".jhipster/modules";
   private static final String HISTORY_FILE = "history.json";
-  private static final String PRESET_FOLDER = "/";
 
   private final ObjectMapper json;
   private final ProjectFormatter formatter;
-  private final ProjectFiles projectFiles;
   private final ObjectWriter writer;
   private final FileSystemProjectDownloader downloader;
-  private final PresetName presetName;
 
-  public FileSystemProjectsRepository(
-    ObjectMapper json,
-    ProjectFormatter formatter,
-    ProjectFiles projectFiles,
-    @Value("${jhlite-preset-file.name:preset.json}") String presetFileName
-  ) {
+  public FileSystemProjectsRepository(ObjectMapper json, ProjectFormatter formatter) {
     this.json = json;
     this.formatter = formatter;
-    this.projectFiles = projectFiles;
-    this.presetName = PresetName.from(presetFileName);
 
     writer = json.writerWithDefaultPrettyPrinter();
     downloader = new FileSystemProjectDownloader();
@@ -98,18 +83,5 @@ class FileSystemProjectsRepository implements ProjectsRepository {
 
   private Path historyFilePath(ProjectPath path) {
     return Paths.get(path.get(), HISTORY_FOLDER, HISTORY_FILE);
-  }
-
-  @Override
-  public Collection<Preset> getPresets() {
-    try {
-      return json.readValue(projectFiles.readBytes(presetFilePath()), PersistedPresets.class).toDomain();
-    } catch (IOException e) {
-      throw GeneratorException.technicalError("Can't read presets: " + e.getMessage(), e);
-    }
-  }
-
-  private String presetFilePath() {
-    return PRESET_FOLDER + presetName.name();
   }
 }
