+++ src/main/java/tech/jhipster/lite/project/infrastructure/secondary/PersistedPreset.java
@@ -1,9 +1,9 @@
 package tech.jhipster.lite.project.infrastructure.secondary;
 
 import java.util.Collection;
+import tech.jhipster.lite.module.domain.preset.Preset;
+import tech.jhipster.lite.module.domain.preset.PresetName;
 import tech.jhipster.lite.project.domain.ModulesSlugs;
-import tech.jhipster.lite.project.domain.preset.Preset;
-import tech.jhipster.lite.project.domain.preset.PresetName;
 
 record PersistedPreset(String name, Collection<String> modules) {
   public Preset toDomain() {
