+++ src/test/java/tech/jhipster/lite/project/infrastructure/secondary/FileSystemProjectsRepositoryTest.java
@@ -24,26 +24,18 @@ import org.junit.jupiter.api.Test;
 import tech.jhipster.lite.JsonHelper;
 import tech.jhipster.lite.TestFileUtils;
 import tech.jhipster.lite.UnitTest;
-import tech.jhipster.lite.module.domain.ProjectFiles;
-import tech.jhipster.lite.project.domain.ModuleSlug;
-import tech.jhipster.lite.project.domain.ModulesSlugs;
 import tech.jhipster.lite.project.domain.ProjectPath;
 import tech.jhipster.lite.project.domain.download.Project;
 import tech.jhipster.lite.project.domain.download.ProjectName;
 import tech.jhipster.lite.project.domain.history.ProjectHistory;
-import tech.jhipster.lite.project.domain.preset.Preset;
-import tech.jhipster.lite.project.domain.preset.PresetName;
 import tech.jhipster.lite.shared.error.domain.GeneratorException;
 
 @UnitTest
 class FileSystemProjectsRepositoryTest {
 
-  private static final String DEFAULT_PRESET_FILE = "preset.json";
   private static final FileSystemProjectsRepository projects = new FileSystemProjectsRepository(
     JsonHelper.jsonMapper(),
-    mock(ProjectFormatter.class),
-    mock(ProjectFiles.class),
-    DEFAULT_PRESET_FILE
+    mock(ProjectFormatter.class)
   );
 
   @Nested
@@ -147,12 +139,7 @@ class FileSystemProjectsRepositoryTest {
       ObjectMapper json = mock(ObjectMapper.class);
       when(json.writerWithDefaultPrettyPrinter()).thenReturn(writer);
 
-      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(
-        json,
-        mock(ProjectFormatter.class),
-        mock(ProjectFiles.class),
-        DEFAULT_PRESET_FILE
-      );
+      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(json, mock(ProjectFormatter.class));
 
       assertThatThrownBy(() -> fileSystemProjectsRepository.save(projectHistory())).isExactlyInstanceOf(GeneratorException.class);
     }
@@ -191,12 +178,7 @@ class FileSystemProjectsRepositoryTest {
       ObjectMapper json = mock(ObjectMapper.class);
       when(json.readValue(any(byte[].class), eq(PersistedProjectHistory.class))).thenThrow(IOException.class);
 
-      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(
-        json,
-        mock(ProjectFormatter.class),
-        mock(ProjectFiles.class),
-        DEFAULT_PRESET_FILE
-      );
+      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(json, mock(ProjectFormatter.class));
 
       assertThatThrownBy(() -> fileSystemProjectsRepository.getHistory(path)).isExactlyInstanceOf(GeneratorException.class);
     }
@@ -219,53 +201,6 @@ class FileSystemProjectsRepositoryTest {
 
       assertThat(history.path()).isEqualTo(path);
       assertThat(history.actions()).usingRecursiveFieldByFieldElementComparator().containsExactly(projectAction());
-    }
-  }
-
-  @Nested
-  @DisplayName("Get preset")
-  class FileSystemProjectsRepositoryGetPresetTest {
-
-    @Test
-    void shouldHandleDeserializationErrors() throws IOException {
-      ObjectMapper json = mock(ObjectMapper.class);
-      when(json.readValue(any(byte[].class), eq(PersistedPresets.class))).thenThrow(IOException.class);
-      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(
-        json,
-        mock(ProjectFormatter.class),
-        mockProjectFilesWithValidPresetJson(),
-        DEFAULT_PRESET_FILE
-      );
-
-      assertThatThrownBy(fileSystemProjectsRepository::getPresets).isExactlyInstanceOf(GeneratorException.class);
-    }
-
-    @Test
-    void shouldNotReturnPresetFromUnknownFile() {
-      ProjectFiles projectFiles = mock(ProjectFiles.class);
-      lenient().when(projectFiles.readBytes("/preset.json")).thenThrow(GeneratorException.class);
-      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(
-        JsonHelper.jsonMapper(),
-        mock(ProjectFormatter.class),
-        projectFiles,
-        DEFAULT_PRESET_FILE
-      );
-
-      assertThatThrownBy(fileSystemProjectsRepository::getPresets).isExactlyInstanceOf(GeneratorException.class);
-    }
-
-    @Test
-    void shouldGetExistingPreset() {
-      FileSystemProjectsRepository fileSystemProjectsRepository = new FileSystemProjectsRepository(
-        JsonHelper.jsonMapper(),
-        mock(ProjectFormatter.class),
-        mockProjectFilesWithValidPresetJson(),
-        DEFAULT_PRESET_FILE
-      );
-
-      Collection<Preset> presets = fileSystemProjectsRepository.getPresets();
-
-      assertThat(presets).containsExactly(expectedPreset());
     }
   }
 
@@ -312,57 +247,5 @@ class FileSystemProjectsRepositoryTest {
     public ProjectPath build() {
       return new ProjectPath(folder.toString());
     }
-  }
-
-  private static ProjectFiles mockProjectFilesWithValidPresetJson() {
-    ProjectFiles projectFiles = mock(ProjectFiles.class);
-
-    String validPresetJson =
-      """
-      {
-        "presets": [
-          {
-            "name": "angular + spring boot",
-            "modules": [
-              "init",
-              "application-service-hexagonal-architecture-documentation",
-              "maven-java",
-              "prettier",
-              "angular-core",
-              "java-base",
-              "maven-wrapper",
-              "spring-boot",
-              "spring-boot-mvc-empty",
-              "logs-spy",
-              "spring-boot-tomcat"
-            ]
-          }
-        ]
-      }
-      """;
-    lenient().when(projectFiles.readBytes("/preset.json")).thenReturn(validPresetJson.getBytes());
-
-    return projectFiles;
-  }
-
-  private static Preset expectedPreset() {
-    return new Preset(
-      new PresetName("angular + spring boot"),
-      new ModulesSlugs(
-        List.of(
-          new ModuleSlug("init"),
-          new ModuleSlug("application-service-hexagonal-architecture-documentation"),
-          new ModuleSlug("maven-java"),
-          new ModuleSlug("prettier"),
-          new ModuleSlug("angular-core"),
-          new ModuleSlug("java-base"),
-          new ModuleSlug("maven-wrapper"),
-          new ModuleSlug("spring-boot"),
-          new ModuleSlug("spring-boot-mvc-empty"),
-          new ModuleSlug("logs-spy"),
-          new ModuleSlug("spring-boot-tomcat")
-        )
-      )
-    );
   }
 }
